requirements.txt
aiohttp==3.7.3
async-timeout==3.0.1
attrs==20.3.0
chardet==3.0.4
discord.py~=1.7.1
dnspython==2.0.0
idna==2.10
motor~=2.4.0
multidict==4.7.6
pymongo~=3.11.2
typing-extensions==3.7.4.3
yarl==1.5.1
aiocache~=0.11.1
aioredis~=1.3.1
asyncio~=3.4.3
loguru~=0.5.1